* src/Timer.cpp, src/gui/Menu.cpp: Correction d'un bug.
git-svn-id: svn://svn.gna.org/svn/menareants/trunk@398 6832cf7c-ce15-0410-9f4e-c008ec2580bf

+++ src/Timer.h
@@ -32,7 +32,7 @@ public:
 
 protected:
 
-	static int begin_time;
+	static long begin_time;
 	static float elapsed;
 };
